Browser Support and Additional Controll Features
This Update resolves Issues on current browsers, and in our tests
Chrome,FF, IE, Opera and Safari are now supported.

Additional Update for Controls, you have now the ability to define if
you want to enable doublezoom and rotate, as well as define the factors
for each of them.
Added working example for minimal controls and updated on page
documentation.